Add files via upload
Intercollegiate Sports Management is a web Application which provides a way for intercollege to maintain sports details. This project is an online working and it’s developed using HTML, CSS, JavaScript and Python , we are developing web application for Intercollegiate Event Management for provide the information about the sports and events like cultural to the students. Now a days online or auto management is very important to save time and man power for handling this job. Each and every people are very closely towards the internet and like a smart work. With the help of website we can easily provide the information about the Sports for the students.